test(reference-video): 拆出独立 integration 文件,避免测试同时继承 unit 与 integration 标记
pytest marker 可加性：给 tmp_path 真实 I/O 的新增测试加 integration 标记不会移除
所在模块 module-level pytestmark = pytest.mark.unit 的继承，该测试仍会被 -m unit
选中。拆分到独立的 integration-only 文件消除这个双重标记。
